refactoring add repos services

+++ web/src/main/java/com/dental/persistence/entity/BaseEntity.java
@@ -5,6 +5,6 @@ package com.dental.persistence.entity;
  */
 public abstract class BaseEntity {
 
-  public abstract Integer getId();
+  public abstract Long getId();
 
 }
